Factor out solver
